app/src/main/java/com/udacity/gradle/builditbigger/MyJokeAsyncTask.java
package com.udacity.gradle.builditbigger;

import android.content.Context;
import android.os.AsyncTask;
import android.util.Pair;

import com.google.api.client.extensions.android.http.AndroidHttp;
import com.google.api.client.extensions.android.json.AndroidJsonFactory;
import com.google.api.client.googleapis.services.AbstractGoogleClientRequest;
import com.google.api.client.googleapis.services.GoogleClientRequestInitializer;
import com.udacity.gradle.builditbigger.backend.myJokeApi.MyJokeApi;
import com.udacity.gradle.builditbigger.backend.myJokeApi.model.MyJoke;

import java.io.IOException;

class MyJokeAsyncTask extends AsyncTask<Pair<Context, String>, Void, String> {
    private static MyJokeApi myApiService = null;
    private JokeListener jokeListener;

    public MyJokeAsyncTask(JokeListener jokeListener) {
        this.jokeListener = jokeListener;
    }

    public MyJokeAsyncTask() {
    }

    public interface JokeListener{
        void JokeReceived(String joke);
    }

    @Override
    protected String doInBackground(Pair<Context, String>... params) {
        if(myApiService == null) {  // Only do this once
//
            MyJokeApi.Builder builder = new MyJokeApi.Builder(AndroidHttp.newCompatibleTransport(),
                    new AndroidJsonFactory(), null)
                    // options for running against local devappserver
                    // - 10.0.2.2 is localhost's IP address in Android emulator
                    // - turn off compression when running against local devappserver
                    .setRootUrl("http://10.0.2.2:8080/_ah/api/")
                    .setGoogleClientRequestInitializer(new GoogleClientRequestInitializer() {
                        @Override
                        public void initialize(AbstractGoogleClientRequest<?> abstractGoogleClientRequest) throws IOException {
                            abstractGoogleClientRequest.setDisableGZipContent(true);
                        }
                    });
            myApiService = builder.build();
        }

        try {
            return myApiService.sayJoke(new MyJoke()).execute().getJoke();
        } catch (IOException e) {
            return e.getMessage();
        }
    }

    @Override
    protected void onPostExecute(String result) {
//        Toast.makeText(context, result, Toast.LENGTH_LONG).show();

        jokeListener.JokeReceived(result);
    }
}